+++ orchestrator_service/src/main/resources/db/changelog/v1.0.0/007-db.changelog-createBugReportTable.xml
@@ -9,7 +9,7 @@
             <column name="id" type="BIGINT" autoIncrement="true">
                 <constraints primaryKey="true" nullable="false"/>
             </column>
-            <column name="report" type="VARCHAR(255)">
+            <column name="report" type="VARCHAR(20000)">
                 <constraints nullable="false"/>
             </column>
             <column name="user_id" type="BIGINT">
