+++ orchestrator_service/src/main/resources/static/scripts/setting_scripts.js
@@ -72,7 +72,7 @@ sendButton.on("click", function() {
     };
 
     $.ajax({
-        url: "/settings/bugreport",
+        url: "/bugreport",
         type: "POST",
         contentType: "application/json",
         data: JSON.stringify(data),
